replicate-requirements.txt
yt-dlp>=2023.3.4
replicate>=0.20.0
tqdm>=4.65.0
requests>=2.28.0
ffmpeg-python>=0.2.0
google-cloud-pubsub>=2.13.0
google-cloud-storage>=2.5.0
google-auth>=2.15.0
python-dotenv>=1.0.0